pyproject.toml
[project]
name = "markitdown-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.38.0",
    "markitdown>=0.1.3",
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
]
